pyproject.toml
[tool.poetry]
name = "principled-prompter"
version = "0.1.0"
description = ""
authors = ["hskimim <hskimim8855@gmail.com>"]
readme = "README.md"
packages = [{include = "principled_prompter"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.3"
pandas = "^2.1.4"
ipykernel = "^6.28.0"
matplotlib = "^3.8.2"
pydantic = "^2.5.3"
openai = "^1.7.2"
python-dotenv = "^1.0.0"
types-tqdm = "^4.66.0.20240106"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
